Moving viewer.js into own service, still buggy
